Compliance Cleanup
Descriptions in Construction.cs, splitting up SurfaceElements into three separate files, one for each return type.
