+++ Structure_Engine/Create/Physical/ISurface.cs
@@ -43,78 +43,6 @@ namespace BH.Engine.Structure
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Creates a physical Wall from a Panel. The Wall will be assigned a Construction based on the SurfaceProperty of the Panel.")]
-        [Input("panel", "The Panel to use as the base for the framing element.")]
-        [Output("wall", "The created physical Wall based on the Panel element provided.")]
-        public static BHPE.Wall Wall(Panel panel)
-        {
-            if (panel.IsNull())
-                return null;
-
-            //Get Construction
-            ISurfaceProperty prop = panel.Property;
-            BHPC.Construction construction = null;
-
-            if (prop == null)
-                Base.Compute.RecordWarning("The panel does not contain a surfaceProperty. Can not extract profile or material");
-            else
-                construction = panel.Property.IConstruction();
-
-            //Get Location
-            PolyCurve externalEdges = Geometry.Create.PolyCurve(panel.ExternalEdges.Select(x => x.Curve));
-            List<PolyCurve> internalEdges = panel.Openings.Select(opening => Geometry.Create.PolyCurve(opening.Edges.Select(edge => edge.Curve))).ToList();
-
-            BHPE.Wall surfaceElement = Physical.Create.Wall(construction, externalEdges, internalEdges);
-
-            string name = panel.Name ?? "";
-            surfaceElement.Name = name;
-
-            if (panel.FindFragment<PanelRebarIntent>() != null || panel.Property.FindFragment<ReinforcementDensity>() != null)
-            {
-                Base.Compute.RecordWarning("The panel has reinforcement, but this is not yet implemented.");
-            }
-
-            return surfaceElement;
-        }
-
-        /***************************************************/
-
-        [Description("Creates a physical Floor from a Panel. The Floor will be assigned a Construction based on the SurfaceProperty of the Panel.")]
-        [Input("panel", "The Panel to use as the base for the framing element.")]
-        [Output("floor", "The created physical Floor based on the Panel element provided.")]
-        public static BHPE.Floor Floor(Panel panel)
-        {
-            if (panel.IsNull())
-                return null;
-
-            //Get Construction
-            ISurfaceProperty prop = panel.Property;
-            BHPC.Construction construction = null;
-
-            if (prop == null)
-                Base.Compute.RecordWarning("The panel does not contain a surfaceProperty. Can not extract profile or material");
-            else
-                construction = panel.Property.IConstruction();
-
-            //Get Location
-            PolyCurve externalEdges = Geometry.Create.PolyCurve(panel.ExternalEdges.Select(x => x.Curve));
-            List<PolyCurve> internalEdges = panel.Openings.Select(opening => Geometry.Create.PolyCurve(opening.Edges.Select(edge => edge.Curve))).ToList();
-
-            BHPE.Floor surfaceElement = Physical.Create.Floor(construction, externalEdges, internalEdges);
-
-            string name = panel.Name ?? "";
-            surfaceElement.Name = name;
-
-            if (panel.FindFragment<PanelRebarIntent>() != null || panel.Property.FindFragment<ReinforcementDensity>() != null)
-            {
-                Base.Compute.RecordWarning("The panel has reinforcement, but this is not yet implemented.");
-            }
-
-            return surfaceElement;
-        }
-
-        /***************************************************/
-
         [Description("Creates a physical surface element from a Panel. The Floor will be assigned a Construction based on the SurfaceProperty of the Panel, and the PanelType of the SufaceProperty will determine what type of surface element to create, unless overridden.")]
         [Input("panel", "The Panel to use as the base for the framing element.")]
         [Input("structuralUsage", "The type of surface element to create. if Undefined, the type will be based on the panel's SurfaceProperty.")]
